CP-314 - Дополнены и изменены таблицы транзакций, параметров смены и GUID параметров. Создан тестовый экран для сохранения транзакций и параметров GUID. Проброшен DI до главного экрана.

+++ persistence/src/main/java/ru/petrolplus/pos/persitence/entities/GUIDparamsDTO.kt
@@ -14,7 +14,7 @@ package ru.petrolplus.pos.persitence.entities
 
 data class GUIDparamsDTO(
     override val id: Int = 1,
-    val lastOnlineTransaction: Int,
+    val lastOnlineTransaction: Long,
     val lastGeneratedTime: Long,
     val clockSequence: Int,
     val hasNodeId: Boolean,
